requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
catboost
xgboost
dilli
-e .